Add #4 : PlainTextJobConfig - 테스트 코드 개발 - 1

+++ src/main/java/com/hyun6ik/batchstudy/core/repository/ResultTextRepository.java
@@ -3,5 +3,5 @@ package com.hyun6ik.batchstudy.core.repository;
 import com.hyun6ik.batchstudy.core.domain.ResultText;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ResultTextRepository extends JpaRepository<ResultText, Long> {
+public interface ResultTextRepository extends JpaRepository<ResultText, Integer> {
 }
